fix(import): allowing user to import 12 words mnemonic
Right now, users can only import a 24-word seed phrase. The process of
converting this phrase back to its original entropy results in a 16-item
array. However, creating the signing key needed to encrypt the seed
phrase requires a 32-item array. To solve this, I currently adds 0s to
the end of the 16-item array until it reaches the required 32 items.